requirements.txt
paho-mqtt>=2.0.0
meshtastic>=2.3.0
pyyaml>=6.0
cryptography>=42.0.0
